+++ central/clusterinit/store/tests/store_test.go
@@ -5,7 +5,10 @@ package tests
 import (
 	"context"
 	"errors"
+	"fmt"
+	"math/rand"
 	"testing"
+	"time"
 
 	"github.com/stackrox/rox/central/clusterinit/store"
 	pgStore "github.com/stackrox/rox/central/clusterinit/store/postgres"
@@ -13,7 +16,9 @@ import (
 	"github.com/stackrox/rox/pkg/postgres"
 	"github.com/stackrox/rox/pkg/postgres/pgtest"
 	"github.com/stackrox/rox/pkg/sac"
+	"github.com/stackrox/rox/pkg/uuid"
 	"github.com/stretchr/testify/suite"
+	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
 func TestClusterInitStore(t *testing.T) {
@@ -117,4 +122,105 @@ func (s *clusterInitStoreTestSuite) TestRevokeToken() {
 	s.Require().NoError(err)
 	s.Equal(toReuseMetaName.GetName(), reused.GetName())
 	s.Equal(toRevokeMeta.GetName(), reused.GetName())
+}
+
+// Tests auto revocation for a CRS with maxRegistrations == 0 (unlimited registrations).
+
+func (s *clusterInitStoreTestSuite) TestCrsWithoutMaxRegistrations() {
+	clusterName := fmt.Sprintf("some-cluster-%s", uuid.NewV4().String())
+	crsId := uuid.NewV4().String()
+	crsMeta := &storage.InitBundleMeta{
+		Id:        crsId,
+		Name:      fmt.Sprintf("test-crs-unlimited-1-%d", rand.Intn(10000)),
+		CreatedAt: timestamppb.New(time.Now()),
+		Version:   storage.InitBundleMeta_CRS,
+	}
+	err := s.store.Add(s.ctx, crsMeta)
+	s.Require().NoError(err, "adding CRS %s failed", crsId)
+
+	err = s.store.InitiateClusterRegistration(s.ctx, crsId, clusterName)
+	s.NoErrorf(err, "recording initiated registration for CRS %s failed", crsId)
+
+	crsMeta, err = s.store.Get(s.ctx, crsMeta.GetId())
+	s.NoErrorf(err, "retrieving CRS %s failed", crsId)
+	s.Empty(crsMeta.GetRegistrationsInitiated(), "CRS %s has registrationsInitiated non-empty, even though registrations are unlimited", crsId)
+
+	err = s.store.MarkClusterRegistrationComplete(s.ctx, crsId, clusterName)
+	s.NoErrorf(err, "recording completed registration for CRS %s failed", crsId)
+	s.Empty(crsMeta.GetRegistrationsInitiated(), "CRS %s has registrationsInitiated non-empty, even though registrations are unlimited", crsId)
+	s.Empty(crsMeta.GetRegistrationsCompleted(), "CRS %s has registrationsCompleted non-empty, even though registrations are unlimited", crsId)
+	s.Falsef(crsMeta.GetIsRevoked(), "CRS %s is revoked", crsId)
+}
+
+// Tests auto revocation for a CRS with maxRegistrations == 1.
+
+func (s *clusterInitStoreTestSuite) TestCrsAutoRevocationOneShot() {
+	clusterName := fmt.Sprintf("some-cluster-%s", uuid.NewV4().String())
+	crsId := uuid.NewV4().String()
+	crsMeta := &storage.InitBundleMeta{
+		Id:               crsId,
+		Name:             fmt.Sprintf("test-crs-auto-revocation-1-%d", rand.Intn(10000)),
+		CreatedAt:        timestamppb.New(time.Now()),
+		Version:          storage.InitBundleMeta_CRS,
+		MaxRegistrations: 1,
+	}
+	err := s.store.Add(s.ctx, crsMeta)
+	s.Require().NoError(err, "adding CRS %s failed", crsId)
+
+	err = s.store.InitiateClusterRegistration(s.ctx, crsId, clusterName)
+	s.NoErrorf(err, "recording initiated registration for CRS %s failed", crsId)
+
+	err = s.store.InitiateClusterRegistration(s.ctx, crsId, clusterName)
+	s.Error(err, "cluster registration still possible")
+
+	err = s.store.MarkClusterRegistrationComplete(s.ctx, crsId, clusterName)
+	s.NoErrorf(err, "recording completed registration for CRS %s failed", crsId)
+
+	crsMeta, err = s.store.Get(s.ctx, crsId)
+	s.NoErrorf(err, "receiving CRS %s", crsId)
+	s.Truef(crsMeta.GetIsRevoked(), "CRS %s is not revoked", crsId)
+
+}
+
+// Tests auto revocation for a CRS with maxRegistrations > 1.
+func (s *clusterInitStoreTestSuite) TestCrsAutoRevocationAfterTwoRegistrations() {
+	clusterName := fmt.Sprintf("some-cluster-%s", uuid.NewV4().String())
+	crsId := uuid.NewV4().String()
+	crsMeta := &storage.InitBundleMeta{
+		Id:               crsId,
+		Name:             fmt.Sprintf("test-crs-auto-revocation-2-%d", rand.Intn(10000)),
+		CreatedAt:        timestamppb.New(time.Now()),
+		Version:          storage.InitBundleMeta_CRS,
+		MaxRegistrations: 2,
+	}
+	err := s.store.Add(s.ctx, crsMeta)
+	s.Require().NoError(err, "adding CRS %s failed", crsId)
+
+	err = s.store.InitiateClusterRegistration(s.ctx, crsId, clusterName)
+	s.NoErrorf(err, "recording initiated registration for CRS %s failed", crsId)
+
+	err = s.store.InitiateClusterRegistration(s.ctx, crsId, clusterName)
+	s.NoError(err, "cluster registration not possible")
+
+	err = s.store.MarkClusterRegistrationComplete(s.ctx, crsId, clusterName)
+	s.NoErrorf(err, "recording completed registration for CRS %s failed", crsId)
+
+	crsMeta, err = s.store.Get(s.ctx, crsId)
+	s.NoErrorf(err, "receiving CRS %s", crsId)
+	s.Falsef(crsMeta.GetIsRevoked(), "CRS %s is not revoked", crsId)
+
+	clusterName = fmt.Sprintf("some-cluster-%s", uuid.NewV4().String())
+
+	err = s.store.InitiateClusterRegistration(s.ctx, crsId, clusterName)
+	s.NoErrorf(err, "recording initiated registration for CRS %s failed", crsId)
+
+	err = s.store.InitiateClusterRegistration(s.ctx, crsId, clusterName)
+	s.Error(err, "cluster registration still possible")
+
+	err = s.store.MarkClusterRegistrationComplete(s.ctx, crsId, clusterName)
+	s.NoErrorf(err, "recording completed registration for CRS %s failed", crsId)
+
+	crsMeta, err = s.store.Get(s.ctx, crsId)
+	s.NoErrorf(err, "receiving CRS %s", crsId)
+	s.Truef(crsMeta.GetIsRevoked(), "CRS %s is not revoked", crsId)
 }